refactor(tui): Move ChannelsView and CostsView to views/ (#1581)
Move view-like components from /components to /views for consistency:

- Move ChannelsView.tsx (503 lines) to views/
- Move CostsView.tsx (225 lines) to views/
- Update imports in app.tsx
- Update imports in test files
- Fix relative imports within moved files

This improves code organization and makes it clear what constitutes
a view (full-page component) vs a component (reusable UI element).
