+++ tui/src/app.tsx
@@ -21,8 +21,8 @@ import { Dashboard } from './views/Dashboard';
 import { AgentsView } from './views/AgentsView';
 import { CommandsView } from './views/CommandsView';
 import { RolesView } from './views/RolesView';
-import { ChannelsView } from './components/ChannelsView';
-import { CostsView } from './components/CostsView';
+import { ChannelsView } from './views/ChannelsView';
+import { CostsView } from './views/CostsView';
 import { LogsView } from './views/LogsView';
 import { WorktreesView } from './views/WorktreesView';
 import { WorkspaceSelectorView } from './views/WorkspaceSelectorView';
